backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for RAG"
authors = ["Shivam Singh"]

[tool.poetry.dependencies]
python = "^3.11.0"
httpx = "^0.27.0"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
qdrant-client = "^1.13.3"
sentence-transformers = "^3.4.1"
requests = "^2.32.3"
PyPDF2 = "^3.0.1"
langchain = "^0.3.21"
langchain-community = "^0.3.20"
langchain-huggingface = "^0.1.2"
langchain-qdrant ="^0.2.0"
huggingface-hub = "^0.29.3"
python-dotenv = "^1.0.1"
PyMuPDF = "^1.25.4"
python-multipart = "^0.0.20"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
